@jesusvazquez 2022-11-28
From a conversation with @bboreham, there is no current benchmark for writting samples so I'll try to write one that benchmarks remote write input which is the way we usually expect out-of-order samples to come from.