@jesusvazquez 2022-11-29
Added first iteration of the benchmark that simply writes ooo samples in order to a thousand different series using the remote write handler which is the way we're expecting ooo samples to come.

Results of benchmark:
```
❯ cat old_implementation.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.09764 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08728 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.09053 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08576 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.09350 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08826 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08777 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.1051 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08660 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08684 ns/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	9.402s
``` 
```
❯ cat new_implementation.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.09207 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08552 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08643 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08470 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08418 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08510 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08482 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08539 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08540 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08579 ns/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	8.366s
```
```
❯ benchstat old_implementation.txt new_implementation.txt
name                     old time/op  new time/op  delta
RemoteWriteOOOSamples-8  0.09ns ± 9%  0.09ns ± 1%  -4.58%  (p=0.000 n=9+9)
``` 

Shows a small improvement which is what we expected.

cc @bboreham please take a look at the benchmark, perhaps I can make it more meaningful.