@jesusvazquez 2022-11-29
I realized that I was reseting the benchmark timer at the wrong place and in the new benchmark I'm doing so after all the requests I'm going to make have been buffered so benchmark time only cares about the writes themselves.

```
❯ cat old_implementation.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05734 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05503 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.06218 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.06008 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.08415 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.06985 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.06145 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.1069 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.07372 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.09939 ns/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	9.973s
```

```
❯ cat new_implementation.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05633 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05306 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05456 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.1290 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05547 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05617 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05513 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05451 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05951 ns/op
BenchmarkRemoteWriteOOOSamples-8   	1000000000	         0.05428 ns/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	7.317s
```

```
❯ benchstat old_implementation.txt new_implementation.txt
name                     old time/op  new time/op  delta
RemoteWriteOOOSamples-8  0.07ns ±46%  0.06ns ± 7%  -24.06%  (p=0.001 n=10+9)
```